added status column
